+++ inbm/dispatcher-agent/dispatcher/aota/constants.py
@@ -31,3 +31,7 @@ CENTOS_DRIVER_PATH = "/host/inb_driver/"
 class SupportedDriver(Enum):
     XLINK = "thb-hddl-xlink-pci-net-driver"
     FLASH_LOGIC = "thb-flash-logic-driver"
+
+# Command Exit Status
+CMD_SUCCESS = 0
+CMD_TERMINATED_BY_SIGTERM = -15