+++ inbm/Changelog.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
  - RTC 530729 - Fix AOTA update log file show Reboot Failed although platform already rebooted and application updated
+ - RTC 530881 - Fix JSON update log/access
 
 ## 4.1.0 - 2023-06-13